Move IGameData to common.
