+++ src/common/game/IGameData.ts
@@ -1,0 +1,7 @@
+import {GameId} from '../Types';
+
+// TODO(kberg): rename to something like IGameCloneMetadata
+export interface IGameData {
+  gameId: GameId;
+  playerCount: number;
+}
